refactor: dedupe useObservable/useSyncObservable internals into shared modules
