+++ packages/react-rx/src/index.ts
@@ -1,5 +1,6 @@
 'use client'
 
+export * from './types'
 export * from './useObservable'
 export * from './useObservableEvent'
 export * from './useSyncObservable'
