+++ packages/react-rx/src/useSyncObservable.ts
@@ -1,55 +1,9 @@
-// NOTE: `useObservable.ts` intentionally duplicates this file (plus a useDeferredValue wrap and a
-// snapshot-based getServerSnapshot). Keep the copies in sync instead of extracting shared helpers
-// or merging the two WeakMap caches; deduplication is a planned follow-up PR.
 import {useCallback, useMemo, useSyncExternalStore} from 'react'
-import {
-  asapScheduler,
-  catchError,
-  finalize,
-  map,
-  type Observable,
-  type ObservedValueOf,
-  of,
-  share,
-  tap,
-  timer,
-} from 'rxjs'
+import {type Observable, type ObservedValueOf} from 'rxjs'
 
-function getValue<T>(value: T): T extends () => infer U ? U : T {
-  return (typeof value === 'function' ? (value as () => any)() : value) as T extends () => infer U
-    ? U
-    : T
-}
-
-interface ObservableState<T> {
-  didEmit: boolean
-  snapshot?: T
-  error?: unknown
-}
-
-interface CacheRecord<T> {
-  observable: Observable<void>
-  state: {
-    didEmit: boolean
-    snapshot?: T
-    error?: unknown
-  }
-  getSnapshot: (initialValue: unknown) => T
-}
-
-const cache = new WeakMap<Observable<any>, CacheRecord<any>>()
-
-const EMPTY_OBJECT = {}
-
-/** @public */
-export interface UseObservableOptions {
-  /**
-   * Pause the active store subscription. While `true`, later emissions do not update the component
-   * and the last received value (or `initialValue`) is returned. Does not skip the render-phase
-   * warm-up subscription — swap the observable if you need zero subscriptions.
-   */
-  disabled?: boolean
-}
+import {getOrCreateStore} from './cache'
+import type {UseObservableOptions} from './types'
+import {EMPTY_OBJECT, getValue} from './utils'
 
 /**
  * Subscribe to an observable and return its latest value synchronously via
@@ -89,66 +43,7 @@ export function useSyncObservable<ObservableType extends Observable<any>, Initia
 ): InitialValue | ObservedValueOf<ObservableType> {
   const {disabled = false} = options
 
-  const instance = useMemo(() => {
-    const cached = cache.get(observable)
-    if (cached) {
-      return cached
-    }
-    // This separate object is used as a stable reference to the cache entry's snapshot and error.
-    // It's used by the `getSnapshot` closure.
-    const state: ObservableState<ObservedValueOf<ObservableType>> = {
-      didEmit: false,
-    }
-    const entry: CacheRecord<ObservedValueOf<ObservableType>> = {
-      state,
-      observable: observable.pipe(
-        map((value) => ({snapshot: value, error: undefined})),
-        catchError((error) => of({snapshot: undefined, error})),
-        tap(({snapshot, error}) => {
-          state.didEmit = true
-          state.snapshot = snapshot
-          state.error = error
-        }),
-        // Note: any value or error emitted by the provided observable will be mapped to the cache entry's mutable state
-        // and the observable is thereafter only used as a notifier to call `onStoreChange`, hence the `void` return type.
-        map((value) => void value),
-        // Ensure that the cache entry is deleted when the observable completes or errors.
-        // Comparing `state` (unique per entry) prevents teardown of a stale, already replaced entry
-        // from deleting its successor.
-        finalize(() => {
-          if (cache.get(observable)?.state === state) {
-            cache.delete(observable)
-          }
-        }),
-        share({resetOnRefCountZero: () => timer(0, asapScheduler)}),
-      ),
-      getSnapshot: (initialValue) => {
-        if (state.error) {
-          throw state.error
-        }
-        return (
-          state.didEmit ? state.snapshot : getValue(initialValue)
-        ) as ObservedValueOf<ObservableType>
-      },
-    }
-
-    // The entry must be added to the cache before the eager subscription below: if the observable
-    // terminates synchronously during it, `finalize` runs right away and must find the entry in order to
-    // delete it. Inserting the entry afterwards would retain it (and its last snapshot or error) for as
-    // long as the source observable lives, and poisoned entries would replay stale errors on remount
-    // instead of re-subscribing the source.
-    cache.set(observable, entry)
-
-    // Eagerly subscribe during render to warm up a synchronous snapshot into `state`. This runs even
-    // when `disabled` is true — `disabled` only pauses the live store subscription below. The
-    // subscribe/unsubscribe here does not keep the observable alive; the store subscription does.
-    const subscription = entry.observable.subscribe()
-    subscription.unsubscribe()
-
-    // For synchronously terminating observables the entry has already been evicted from the cache again
-    // at this point, so return the local reference instead of reading back from the cache.
-    return entry
-  }, [observable])
+  const instance = useMemo(() => getOrCreateStore(observable), [observable])
 
   const subscribe = useCallback(
     (onStoreChange: () => void) => {
@@ -169,6 +64,8 @@ export function useSyncObservable<ObservableType extends Observable<any>, Initia
     () => {
       return instance.getSnapshot(initialValue)
     },
+    // Strict v4 server contract: the server renders the resolved `initialValue`, and throws
+    // (missing getServerSnapshot) without one — even when the observable emits synchronously.
     typeof initialValue === 'undefined'
       ? undefined
       : () => getValue(initialValue) as ObservedValueOf<ObservableType>,
